[FP-66] Rename: 경로 변경 및 파일 이동
